+++ src/Navigation.php
@@ -13,11 +13,13 @@ use craft\base\Plugin;
 use craft\events\ConfigEvent;
 use craft\events\RegisterComponentTypesEvent;
 use craft\events\RegisterUrlRulesEvent;
+use craft\events\RegisterUserPermissionsEvent;
 use craft\helpers\UrlHelper;
 use craft\services\Elements;
 use craft\services\Fields;
 use craft\services\ProjectConfig;
 use craft\services\Sites;
+use craft\services\UserPermissions;
 use craft\web\UrlManager;
 use craft\web\twig\variables\CraftVariable;
 
@@ -29,7 +31,7 @@ class Navigation extends Plugin
     // Public Properties
     // =========================================================================
 
-    public $schemaVersion = '1.0.7';
+    public $schemaVersion = '1.0.9';
     public $hasCpSettings = true;
     public $hasCpSection = true;
 
@@ -57,6 +59,7 @@ class Navigation extends Plugin
         $this->_registerProjectConfigEventListeners();
         $this->_registerFieldTypes();
         $this->_registerElementTypes();
+        $this->_registerPermissions();
     }
 
     public function getPluginName()
@@ -140,6 +143,21 @@ class Navigation extends Plugin
     {
         Event::on(Elements::class, Elements::EVENT_REGISTER_ELEMENT_TYPES, function(RegisterComponentTypesEvent $event) {
             $event->types[] = Node::class;
+        });
+    }
+
+    private function _registerPermissions()
+    {
+        Event::on(UserPermissions::class, UserPermissions::EVENT_REGISTER_PERMISSIONS, function(RegisterUserPermissionsEvent $event) {
+            $navs = $this->getNavs()->getAllNavs();
+
+            $navPermissions = [];
+
+            foreach ($navs as $nav) {
+                $navPermissions['navigation-manageNav:' . $nav->uid] = ['label' => Craft::t('navigation', 'Manage “{type}”', ['type' => $nav->name])];
+            }
+
+            $event->permissions[Craft::t('navigation', 'Navigation')] = $navPermissions;
         });
     }
 }