+++ src/models/Nav.php
@@ -1,8 +1,11 @@
 <?php
 namespace verbb\navigation\models;
 
+use verbb\navigation\records\Nav as NavRecord;
+
 use craft\base\Model;
 use craft\validators\HandleValidator;
+use craft\validators\UniqueValidator;
 
 class Nav extends Model
 {
@@ -23,11 +26,17 @@ class Nav extends Model
     // Public Methods
     // =========================================================================
 
+    public function __toString()
+    {
+        return $this->handle;
+    }
+
     public function rules()
     {
         return [
             [['id', 'structureId', 'maxLevels'], 'number', 'integerOnly' => true],
             [['handle'], HandleValidator::class, 'reservedWords' => ['id', 'dateCreated', 'dateUpdated', 'uid', 'title']],
+            [['handle'], UniqueValidator::class, 'targetClass' => NavRecord::class],
             [['name', 'handle'], 'required'],
             [['name', 'handle'], 'string', 'max' => 255],
         ];